Oppsummering av samværsavtale per barn skal opptre under svar på vilkår om aleneomsorg
