+++ src/server/blankett/components/AleneomsorgGrunnlag.tsx
@@ -1,20 +1,12 @@
 import React from 'react';
 import { AnnenForelder } from './AnnenForelder';
-import {
-  IBarnMedSamvær,
-  IPersonalia,
-  samværsandelTilVerdi,
-  Samværsavtale,
-  Samværsdag,
-  Samværsuke,
-} from '../../../typer/dokumentApiBlankett';
+import { IBarnMedSamvær, IPersonalia } from '../../../typer/dokumentApiBlankett';
 import { formaterNullableIsoDato } from '../../utils/util';
 
 interface Props {
   barnMedSamvær: IBarnMedSamvær[];
   personalia: IPersonalia;
   barnId?: string;
-  samværsavtaler?: Samværsavtale[]; //TODO: Fjern nullable etter at familie-ef-sak har blitt prodsatt
 }
 
 const navnPåBarnet = (barnMedSamvær: IBarnMedSamvær) => {
@@ -38,34 +30,7 @@ const utledBostedTekst = (harDeltBosted: boolean, harSammeAdresse: boolean | und
   return 'Ikke registrert på brukers adresse';
 };
 
-export const utledVisningstekstForSamværsavtale = (samværsuker: Samværsuke[]) => {
-  const summertSamvær = samværsuker
-    .flatMap(samværsuke =>
-      Object.values(samværsuke).flatMap((samværsdag: Samværsdag) => samværsdag.andeler),
-    )
-    .map(andel => samværsandelTilVerdi[andel])
-    .reduce((acc, andel) => acc + andel, 0);
-
-  const maksimalSamværsandel = samværsuker.length * 7 * 8;
-
-  const antallHeleDagerMedSamvær = Math.floor(summertSamvær / 8);
-
-  const rest = summertSamvær % 8;
-  const restSuffix = rest === 0 ? '' : '/8';
-
-  const prosentandel = summertSamvær / maksimalSamværsandel;
-
-  const visningstekstAntallDager = `${antallHeleDagerMedSamvær} dager og ${rest}${restSuffix} deler`;
-  const visningstekstProsentandel = `${Math.round(prosentandel * 1000) / 10}%`;
-  return `Samværsandel: ${visningstekstAntallDager} = ${visningstekstProsentandel}`;
-};
-
-export const AleneomsorgGrunnlag: React.FC<Props> = ({
-  barnMedSamvær,
-  barnId,
-  personalia,
-  samværsavtaler,
-}) => {
+export const AleneomsorgGrunnlag: React.FC<Props> = ({ barnMedSamvær, barnId, personalia }) => {
   return (
     <>
       <h3 className={'blankett'}>Registerdata</h3>
@@ -76,10 +41,6 @@ export const AleneomsorgGrunnlag: React.FC<Props> = ({
             barn.registergrunnlag;
 
           const skalViseAdresser = !harDeltBostedVedGrunnlagsdataopprettelse && !harSammeAdresse;
-
-          const samværsavtale = samværsavtaler?.find(
-            samværsavtale => samværsavtale.behandlingBarnId === barnId,
-          );
 
           return (
             <div key={index}>
@@ -105,7 +66,6 @@ export const AleneomsorgGrunnlag: React.FC<Props> = ({
                 {barn.registergrunnlag?.forelder?.visningsadresse ||
                   'Mangler gjeldende bostedsadresse'}
               </div>
-              {samværsavtale && <div>{utledVisningstekstForSamværsavtale(samværsavtale.uker)}</div>}
             </div>
           );
         })}
